refactor: decouple popular networks to prevent e2e cjs import issues
